refactor(review-workflows): refactor middleware
